refine the code structure
