+++ GenAIComps/embeddings/requirements.txt
