+++ GenAIComps/proto/docarray.py
@@ -11,3 +11,25 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+from docarray import BaseDoc, DocList
+from docarray.typing import NdArray
+
+
+class TextDoc(BaseDoc):
+    text: str
+
+
+class EmbedDoc768(BaseDoc):
+    text: str
+    embedding: NdArray[768]
+
+
+class EmbedDoc1024(BaseDoc):
+    text: str
+    embedding: NdArray[1024]
+
+
+class GenerateDoc(BaseDoc):
+    text: str
+    prompt: str
